+++ data/macros/light-tracker-ui-toggle.js
@@ -1,0 +1,10 @@
+/**
+ ****************************************************************
+ * This macro can be used to toggle the Light Tracker interface
+ * on and off.
+ *
+ * NOTE: Only users with the Game Master user role can run this
+ * macro.
+ ***************************************************************/
+
+game.shadowdark.lightSourceTracker.toggleInterface();