+++ system/src/dice/RollSD.mjs
@@ -310,7 +310,7 @@ export default class RollSD extends Roll {
 
 		// Improve the base damage die if this weapon has the relevant property
 		const weaponDamageDieImprovementByProperty =
-			data.actor.system.bonuses.weaponDamageDieImprovementByProperty;
+			data.actor.system.bonuses.weaponDamageDieImprovementByProperty ?? [];
 
 		for (const property of weaponDamageDieImprovementByProperty) {
 			if (await data.item.hasProperty(property)) {
@@ -319,10 +319,12 @@ export default class RollSD extends Roll {
 					shadowdark.config.DAMAGE_DICE
 				);
 
-				versatileDamageDie = shadowdark.utils.getNextDieInList(
-					versatileDamageDie,
-					shadowdark.config.DAMAGE_DICE
-				);
+				if (versatileDamageDie) {
+					versatileDamageDie = shadowdark.utils.getNextDieInList(
+						versatileDamageDie,
+						shadowdark.config.DAMAGE_DICE
+					);
+				}
 			}
 		}
 
@@ -331,7 +333,7 @@ export default class RollSD extends Roll {
 			t => [data.item.name.slugify(), data.item.system.baseWeapon.slugify()].includes(t)
 		)) {
 			damageDie = "d12";
-			versatileDamageDie = "d12";
+			if (versatileDamageDie) versatileDamageDie = "d12";
 		}
 
 		// Check and handle critical failure/success